refactor(cli): start ocean theme — migrate argo + nwm onto discovery
Set up the ocean earthlens.cli entry point (earthlens._ocean_cli) and migrate
argo (validator) and the full nwm cluster. nwm's refresh + validate-live helpers
that were split across core's refresh.py/validate.py are unified into one
earthlens.nwm.cli module (bucket primitives, refresher, curated_ids, validator,
live_validator). Relocate nwm's unit tests and its live e2e to earthlens-ocean.
Shrink the guard allow-lists (argo, nwm; nwm import).
