+++ libs/core/src/earthlens/cli/refresh.py
@@ -1129,163 +1129,6 @@ def _jaxa_grouped(catalog: Any) -> dict[str, list[str]]:
     }
 
 
-#: Region the unsigned NWM operational bucket is listed from.
-_NWM_REGION = "us-east-1"
-
-#: A trailing `_mem<N>` ensemble-member directory suffix (e.g.
-#: `medium_range_mem3`) — collapsed back to its base configuration key.
-_NWM_MEMBER_RE = re.compile(r"_mem\d+$")
-
-
-def _nwm_collapse_member(directory: str) -> str:
-    """Collapse an NWM ensemble-member directory to its configuration key.
-
-    The operational bucket publishes each ensemble member under its own
-    `{config}_mem<N>` directory; the curated catalog keys an ensemble by its
-    bare `{config}` name. Stripping the `_mem<N>` suffix maps a live
-    directory back into the curated namespace so the diff lines up.
-
-    Args:
-        directory: A live configuration directory name.
-
-    Returns:
-        The base configuration key (unchanged when not a member directory).
-
-    Examples:
-        - An ensemble member collapses to its base configuration:
-
-            ```python
-            >>> from earthlens.cli.refresh import _nwm_collapse_member
-            >>> _nwm_collapse_member("medium_range_mem3")
-            'medium_range'
-
-            ```
-        - A non-ensemble directory is returned unchanged:
-
-            ```python
-            >>> from earthlens.cli.refresh import _nwm_collapse_member
-            >>> _nwm_collapse_member("short_range")
-            'short_range'
-
-            ```
-    """
-    return _NWM_MEMBER_RE.sub("", directory)
-
-
-def _nwm_unsigned_client() -> Any:
-    """Return an unsigned `boto3` S3 client for the public `noaa-nwm-pds` bucket.
-
-    The single place the NWM CLI tooling (`refresh` here, `validate --live` in
-    :mod:`earthlens.cli.validate`) builds its anonymous client, so the region
-    and signature config live in one home.
-
-    Returns:
-        An unsigned `boto3` S3 client.
-    """
-    import boto3
-    from botocore import UNSIGNED
-    from botocore.client import Config
-
-    return boto3.client(
-        "s3", region_name=_NWM_REGION, config=Config(signature_version=UNSIGNED)
-    )
-
-
-def _nwm_latest_complete_day(client: Any) -> str:
-    """Return the most recent *complete* `nwm.YYYYMMDD/` day prefix.
-
-    Lists the `nwm.YYYYMMDD/` date prefixes and returns the day before the
-    latest (the newest prefix may be mid-publication), falling back to the
-    only day when just one is present. Shared by the NWM refresh and
-    `validate --live` walks so the "complete day" heuristic lives in one home.
-
-    Args:
-        client: An unsigned S3 client (see :func:`_nwm_unsigned_client`).
-
-    Returns:
-        The selected `nwm.YYYYMMDD` prefix (no trailing slash).
-
-    Raises:
-        RuntimeError: If the bucket exposes no `nwm.YYYYMMDD/` date prefix.
-    """
-    from earthlens.nwm import BUCKET
-
-    paginator = client.get_paginator("list_objects_v2")
-    days = sorted(
-        prefix.rstrip("/")
-        for page in paginator.paginate(Bucket=BUCKET, Delimiter="/")
-        for entry in page.get("CommonPrefixes", [])
-        if (prefix := entry["Prefix"]).startswith("nwm.")
-    )
-    if not days:
-        raise RuntimeError(f"no nwm.YYYYMMDD/ prefixes found on {BUCKET}")
-    return cast("str", days[-2] if len(days) > 1 else days[-1])
-
-
-def _nwm_config_dirs(client: Any, day: str) -> list[str]:
-    """Return the configuration directory names published under one NWM day.
-
-    Args:
-        client: An unsigned S3 client (see :func:`_nwm_unsigned_client`).
-        day: An `nwm.YYYYMMDD` prefix (see :func:`_nwm_latest_complete_day`).
-
-    Returns:
-        The raw configuration directory names (ensemble members still carry
-        their `_mem<N>` suffix), sorted.
-    """
-    from earthlens.nwm import BUCKET
-
-    result = client.list_objects_v2(Bucket=BUCKET, Prefix=f"{day}/", Delimiter="/")
-    return sorted(
-        entry["Prefix"].split("/")[1] for entry in result.get("CommonPrefixes", [])
-    )
-
-
-def _nwm_live_config_dirs() -> list[str]:
-    """Return the configuration directories under the latest complete NWM day.
-
-    Composes the shared NWM bucket primitives: an unsigned client
-    (:func:`_nwm_unsigned_client`), the most recent complete day
-    (:func:`_nwm_latest_complete_day`), and that day's configuration
-    directories (:func:`_nwm_config_dirs`).
-
-    Returns:
-        The raw configuration directory names (ensemble members still carry
-        their `_mem<N>` suffix).
-
-    Raises:
-        RuntimeError: If the bucket exposes no `nwm.YYYYMMDD/` date prefix.
-    """
-    client = _nwm_unsigned_client()
-    return _nwm_config_dirs(client, _nwm_latest_complete_day(client))
-
-
-def _nwm_grouped(catalog: Any) -> dict[str, list[str]]:
-    """List the live NWM configurations from the unsigned operational bucket.
-
-    Walks the most recent complete `nwm.YYYYMMDD/` day on `noaa-nwm-pds`
-    and collapses each ensemble-member directory to its base configuration
-    key (see :func:`_nwm_collapse_member`), so the live set is diffed
-    against the catalog's `available_configurations:` index in the same
-    namespace. The assimilation-input `usgs_timeslices` directory is a live
-    configuration the catalog deliberately does not curate, so it surfaces
-    as an untracked id (informational), not drift.
-
-    Args:
-        catalog: The loaded NWM `Catalog` (unused; the bucket is the source).
-
-    Returns:
-        A single-group mapping `{"nwm": [sorted configuration keys]}`.
-    """
-    dirs = {_nwm_collapse_member(name) for name in _nwm_live_config_dirs()}
-    return {"nwm": sorted(dirs)}
-
-
-def _nwm_curated_configs(catalog: Any) -> list[str]:
-    """Return the configuration keys the NWM catalog curates (its refresh axis)."""
-    return sorted(catalog.configurations)
-
-
 def _s3_grouped(catalog: Any) -> dict[str, list[str]]:
     """List the S3 registry's dataset names (its `available_datasets` universe).
 
@@ -1702,7 +1545,6 @@ _REFRESHERS: dict[str, Callable[[Any], dict[str, list[str]]]] = {
     "radar": _radar_grouped,
     "chc": _chc_grouped,
     "s3": _s3_grouped,
-    "nwm": _nwm_grouped,
     "jaxa": _jaxa_grouped,
     "erddap": _erddap_grouped,
     "gbif": _gbif_grouped,
@@ -1941,7 +1783,6 @@ _CURATED_IDS: dict[str, Callable[[Any], list[str]]] = {
     "worldpop": _worldpop_curated_ids,
     "usgs_water": _curated_attr_ids("code"),
     "chc": _chc_ftp_bases,
-    "nwm": _nwm_curated_configs,
     "gbif": _biodiversity_curated_ids,
     "obis": _biodiversity_curated_ids,
     "wdpa": _biodiversity_curated_ids,
@@ -1954,7 +1795,6 @@ _CURATED_IDS: dict[str, Callable[[Any], list[str]]] = {
 _INDEX_ATTR: dict[str, str] = {
     # Discovered config first; in-core literals are the migration remainder.
     **config_table("index_attr"),
-    "nwm": "available_configurations",
 }
 
 #: Provider id -> a callable computing the bundled ids to diff against, for
